# specialize `isassigned` for performance